break out code into weather and psychro files
