+++ .gitignore
@@ -1,6 +1,7 @@
 OLD
 Debug
 weather
+tmy3
 rc
 out
 *.o
