+++ functions.h
@@ -7,8 +7,6 @@
 
 using namespace std;
 
-const double C_TO_K = 273.15;
-
 // additional data types
 struct atticVent_struct {
 	int wall;
@@ -61,19 +59,6 @@ struct flue_struct {
 	double flueC;
 	double flueHeight;
 	double flueTemp;
-};
-
-struct weatherData {
-	int directNormal;				// Direct normal irradiance (Wh/m2)
-	int globalHorizontal;		// Global horizontal irradiance (Wh/m2)
-	double dryBulb;				// Dry-bulb temperature (deg K)
-	double dewPoint;				// Dew-point temperature (deg K)
-	double relativeHumidity;	// Relative humidity (%)
-	double humidityRatio;		// Humidity Ratio (g/g)
-	double windSpeed;				// Wind speed (m/s)
-	int windDirection;			// wind direction (degrees from North)
-	int pressure;					// Station pressure (Pa)
-	double skyCover;				// Total sky cover (fraction)
 };
 
 // Additional functions
